requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-SQLAlchemy==1.0
Frozen-Flask==0.11
Jinja2==2.7.2
MarkupSafe==0.19
Pillow==2.3.1
PyYAML==3.11
SQLAlchemy==0.9.3
Unidecode==0.04.14
Werkzeug==0.9.4
cssmin==0.2.0
itsdangerous==0.23
jsmin==2.0.9
python-slugify==0.0.7
webassets==0.9
wsgiref==0.1.2
